Default to an empty entity when the node cannot be found.
This allows the class_mapping functionality to run on nodes that don't
have (or need) their own node configuration file. So for example, if a
node isn't explicitly configured and you have a 'default' rule in your
class mapping, then it will run for all nodes, not just those that are
explicitly configured.